refactor: Combine gateway with voice
