--- src/main/java/com/example/pitching/gateway/event/GatewayEvent.java
@@ -1,4 +1,0 @@
-package com.example.pitching.gateway.event;
-
-public interface GatewayEvent {
-}
