+++ src/main/java/com/example/pitching/voice/config/HandlerMappingConfig.java
@@ -1,7 +1,7 @@
-package com.example.pitching.gateway.config;
+package com.example.pitching.voice.config;
 
-import com.example.pitching.gateway.dto.properties.ServerProperties;
-import com.example.pitching.gateway.handler.GatewayWebSocketHandler;
+import com.example.pitching.voice.dto.properties.ServerProperties;
+import com.example.pitching.voice.handler.GatewayWebSocketHandler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
