+++ src/main/java/com/example/pitching/voice/dto/properties/ServerProperties.java
@@ -1,4 +1,4 @@
-package com.example.pitching.gateway.dto.properties;
+package com.example.pitching.voice.dto.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.web.util.UriComponentsBuilder;
