+++ src/main/java/com/example/pitching/voice/event/GatewayEvent.java
@@ -1,0 +1,4 @@
+package com.example.pitching.voice.event;
+
+public interface GatewayEvent {
+}
