+++ src/main/java/com/example/pitching/voice/event/IdentifyEvent.java
@@ -1,7 +1,7 @@
-package com.example.pitching.gateway.event;
+package com.example.pitching.voice.event;
 
-import com.example.pitching.gateway.event.data.IdentifyData;
-import com.example.pitching.gateway.event.op.ReqOp;
+import com.example.pitching.voice.event.data.IdentifyData;
+import com.example.pitching.voice.event.op.ReqOp;
 
 public record IdentifyEvent(
         ReqOp reqOp,
