+++ src/main/java/com/example/pitching/voice/event/data/HelloData.java
@@ -1,4 +1,4 @@
-package com.example.pitching.gateway.event.data;
+package com.example.pitching.voice.event.data;
 
 public record HelloData(long heartbeatInterval) {
     public static HelloData of(long interval) {
