+++ src/main/java/com/example/pitching/voice/event/data/ReadyData.java
@@ -1,4 +1,4 @@
-package com.example.pitching.gateway.event.data;
+package com.example.pitching.voice.event.data;
 
 import java.util.List;
